% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_bunnies.R
\name{add_bunnies}
\alias{+.dust}
\alias{add_bunnies}
\title{Modify Attributes of a \code{dustpan} table}
\usage{
\method{+}{dust}(x, y)
}
\arguments{
\item{x}{a \code{dust} object}

\item{y}{a \code{dust_bunny} object.}
}
\description{
Adds attributes of a plot to be rendered by the \code{print.dust} method.
}
\examples{
x <- dust(lm(mpg ~ qsec + factor(am), data = mtcars))
x
x + dust_colnames(term = "Term", statistic = "T")
x + dust_colnames("Term", "Estimate", "SE", "T-statistic", "p-value")
\dontrun{
# Causes an error due to too few unnamed arguments
x + dust_colnames("Term", "Estimate")
}
}
\author{
Benjamin Nutter
}

