% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_sprinkles.R
\name{add_sprinkles}
\alias{+.dust}
\alias{add_sprinkles}
\title{Modify Attributes of a \code{dustpan} table}
\usage{
\method{+}{dust}(x, y)
}
\arguments{
\item{x}{a \code{dust} object}

\item{y}{a \code{dust_bunny} object.}
}
\description{
Adds attributes of a plot to be rendered by the \code{print.dust} method.
}
\author{
Benjamin Nutter
}
\seealso{
\code{\link{dust}}, \code{\link{sprinkle}}
}

