% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{\%>\%}
\alias{\%>\%}
\title{magrittr forward-pipe operator}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs, rhs}{A dataset and function to apply to it}
}
\description{
Pipe an object forward into a function or call expression
}
