% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Chain together multiple operations}
\usage{
lhs \%<>\% rhs
}
\arguments{
\item{lhs, rhs}{A data set and function to apply it to}
}
\description{
Chain together multiple operations and save to the object
  at the start of the chain.  See `magrittr` documentation for details.
}
