% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dust.R
\name{dust}
\alias{dust}
\title{Prepare an Object for Printing}
\usage{
dust(object, ...)
}
\arguments{
\item{object}{An object that has a \code{tidy} method in \code{broom}}

\item{...}{Additional arguments to pass to \code{tidy}}
}
\value{
Returns an object of class \code{dust}
}
\description{
Generates the basic structure from which \code{dustpan} tables are formatted
  for printing.
}
\details{
\code{dust} is intended to work directly on objects, although matrices and data frames
  may be printed as well.  Formatting of the table may be done by adding additional layers.
}
\examples{
x <- dust(lm(mpg ~ qsec + factor(am), data = mtcars))
x
}
\author{
Benjamin Nutter
}
\seealso{
\link[broom]{tidy}
}

