% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dust_inline.lm.R
\name{dust_inline.lm}
\alias{dust_inline.lm}
\title{Inline Summary for \code{lm} coefficients}
\usage{
\method{dust_inline}{lm}(object, term, level = NULL, se = FALSE,
  test_stat = FALSE, conf_int = TRUE, p_value = TRUE,
  label = "$\\\\beta$", fun = identity, fun_arg = list(), alpha = 0.05,
  ..., style = "none")
}
\arguments{
\item{object}{An object that may be tidied.}

\item{term}{A character vector naming the term (or combination of terms)
in the model for which the summary is desired.  When \code{term} has 
length 1, a main effect is returned.  When the length is greater than 
1, the interaction term between those effects is returned.}

\item{level}{A vector or list, optionally named, that gives the character 
values of the levels for which coefficients are desired when 
\code{term} is a factor.}

\item{se}{\code{logical(1)}. Should the standard error of the estimate be
included in the summary?}

\item{test_stat}{\code{logical(1)}. Should the test statistic for the 
estimate be included in the summary?}

\item{conf_int}{\code{logical(1)}. Should the confidence interval be 
included in the summary?}

\item{p_value}{\code{logical(1)}. Should the p-value for the estimate 
be included in the summary?}

\item{label}{\code{character(1)}. The character label for the term. For 
example, \code{"Estimate"}, \code{"OR"}, or \code{"HR"}. Defaults to
\code{"$\\\\beta$"}, for something generic.}

\item{fun}{A function to apply to the estimate and confidence intervals.
Useful if the response in \code{object} has been transformed. Then 
\code{fun} should be the inverse function of the response.}

\item{fun_arg}{A list of additional arguments for \code{fun}}

\item{alpha}{Significance level (or 1 - confidence) for the confidence 
interval}

\item{...}{Additional arguments to pass to other methods. Not used.}

\item{style}{A \code{character(1)} specifying the output style.  When not
\code{"none"}, this overrules any arguments defined in the call.}
}
\description{
Generates an inline summary of a linear model coefficient.
}
\section{}{
 Functional Requirements
In addition to the requirements listed here, the requirements listed under
\code{\link{dust_inline}} must be satisfied.

\enumerate{
  \item Accepts a character vector naming the \code{term} to be summarised. 
    A length one vector returns the main effect. 
    A length two vector returns the interaction between two terms, etc.
  \item Return an error if no term exists that satisfies the linear 
    combination.
  \item Accepts a vector or list of characters, optionally named, 
    specifying the level for any factors named in term. 
    If unnamed, the levels are assumed to follow the same order of 
    factors in term.
  \item Returns an error if any levels in level cannot be found 
    in its corresponding term.
  \item Accept a \code{logical(1)} indicating if the confidence interval 
    is to be included in the summary
  \item Accept a \code{logical(1)} indicating if the SE is to be included in 
    the summary
  \item Accept a \code{logical(1)} indicating if the test statistic is to be 
    included in the summary
  \item Accept a \code{logical(1)} indicating if the p-value is to be 
    included in the summary
  \item Accept a \code{character(1)} designating the text label for the 
    coefficient (beta, OR, etc)
  \item Accept a function by to apply to the coefficient and CI
  \item Accept additional arguments to the transformation function
}
}
\seealso{
\code{link{dust_inline}}
}

