% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dust_inline.R
\name{dust_inline}
\alias{dust_inline}
\title{Inline Test and Coefficient Summaries}
\usage{
dust_inline(object, ..., style = "none")
}
\arguments{
\item{object}{An object that has a \code{tidy} method.}

\item{...}{Additional arguments to pass to other methods.}

\item{style}{\code{character(1)} specifying the output style.  The 
default value (\code{"none"}) uses other arguments in the method
to determine the output.  Any other value will override the 
other arguments.}
}
\description{
Quickly generate formatted inline summaries for hypothesis 
  tests and model coefficients.
}
\section{Functional Requirements}{

Functional requirements listed here will apply to all \code{dust_inline} 
methods.  Requirements to each method will be documented separately.

\enumerate{
  \item Accepts an object that may be successfully tidyied.
  \item Returns the error message from \code{tidy} when \code{tidy} 
    is not successful
  \item Returns any warnings generated by \code{tidy}
  \item Accepts a \code{character(1)} argument that can determine the 
    output format (overriding other formal arguments)
}
}
\seealso{
\code{\link{dust_inline.lm}}
}

