% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dust_inline.lm.R
\name{dust_inline_unexported}
\alias{attempt_to_tidy}
\alias{dust_inline_unexported}
\alias{force_named_level}
\title{Unexported Utilities for \code{dust_inline}}
\usage{
attempt_to_tidy(object, coll, conf_int, alpha)

force_named_level(object, level, coll)
}
\arguments{
\item{object}{An object to be tidied}

\item{coll}{An \code{AssertCollection} object.}

\item{conf_int}{The value of \code{conf_int} from \code{dust_inline.lm}}

\item{alpha}{The value of \code{alpha} from \code{dust_inline.lm}}

\item{level}{A character vector of levels to match to terms.}
}
\description{
Utilites for the \code{dust_inline} method. These are 
  unexported functions that do nothing but simplify the code, and 
  make reasoning the various parts a little simpler.
}
\details{
\code{attempt_to_tidy} returns a tidied object, if it can, or
  the error generated upon failure.
}

