% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dust_print_method.R
\name{dust_print_method}
\alias{dust_print_method}
\title{Set the Print Method for a \code{dust} Table}
\usage{
dust_print_method(print_method)
}
\arguments{
\item{print_method}{A character string denoting the printing method for the table.  This
  must have length 1.}
}
\description{
Set the Print Method for a \code{dust} Table
}
\details{
\code{dust} supports printing to the console, markdown, html, and latex.  By default,
  tables are printed to the console, unless the user changes \code{options()$dust_format}.
  Printing methods may be changed in any dust call by invoking \code{dust_print_format}.
}
\examples{
x <- dust(lm(mpg ~ qsec + factor(am), data = mtcars))

x$print_method

x <- x + dust_print_method("html")
x$print_method

x <- x + dust_print_method("markdown")
x$print_method
}
\author{
Benjamin Nutter
}

