% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid_color.R
\name{is_valid_color}
\alias{is_valid_color}
\alias{is_valid_color_single}
\title{Test a Character String For Pixiedust Recognized Color Format}
\usage{
is_valid_color(color)

is_valid_color_single(color)
}
\arguments{
\item{color}{A character vector of color names.}
}
\description{
\code{pixiedust} recognizes colors as dvips names, 
  \code{rgb(R,G,B)}, \code{rgba(R,G,B,A)}, \code{#RRGGBB}, or 
  \code{#RRGGBBAA}.  This code returns a logical indicating if 
  the given character strings are valid.
}
\section{Functional Requirements}{

\enumerate{
 \item Returns a logical vector correctly identifying valid color formats.
 \item Casts an error if \code{color} is not a character object.
}
}

