% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.dust.R
\name{knit_print.dust}
\alias{knit_print.dust}
\alias{knit_print.dust_list}
\title{\code{knitr} Printing Function}
\usage{
\method{knit_print}{dust}(x, options, ...)

\method{knit_print}{dust_list}(x, options, ...)
}
\arguments{
\item{x}{A dust object}

\item{options}{A list of options received from the chunk options.}

\item{...}{Additional arguments to pass to other methods.}
}
\description{
Custom printing functions for displaying \code{dust} and 
  \code{dust_list} objects in R Markdown documents.
}
