% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.dust.R
\name{print.dust}
\alias{print.dust}
\title{Print A \code{dust} Table}
\usage{
\method{print}{dust}(x, ..., asis = TRUE)
}
\arguments{
\item{x}{An object of class \code{dust}}

\item{...}{Additional arguments to pass to the print method.  Currently ignored.}

\item{asis}{A logical value that controls if the output is printed using
  \code{knitr::asis_output}.}
}
\description{
Print A \code{dust} Table
}
\details{
Apply the formatting to a \code{dust} object and print the table.
}
\examples{
dust(lm(mpg ~ qsec + factor(am), data = mtcars))
}
\author{
Benjamin Nutter
}

