% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_internal.R
\name{rbind_internal}
\alias{rbind_internal}
\alias{.rbind_internal}
\title{Bind Rows in Base R}
\usage{
.rbind_internal(..., deparse.level = 1)
}
\arguments{
\item{...}{data frames}

\item{deparse.level}{See \code{deparse.level} in \code{rbind}.}
}
\description{
Stack data frames on top of each other. Data frames do not have to 
  have all of the same columns.
}
\author{
Benjamin Nutter
}
