% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_data_internal.R
\name{reshape_data_internal}
\alias{reshape_data_internal}
\alias{.make_dataframe_long}
\title{Reshape data frames for Pixiedust}
\usage{
.make_dataframe_long(data)
}
\arguments{
\item{data}{A \code{data.frame}}
}
\description{
Pixiedust reshapes data to have one row per cell in the table. This 
  permits adjustments to be made to individual cells.  These internal functions
  are provided to simplify the reshaping process. It is slower than using 
  the tidyr functions `gather` and `spread` (or whatever their newer counterparts 
  are), but keeps me off of other people's development schedules.
}
\details{
No validations are performed in these functions, and it is assumed that
  the input data set has the components it needs.
}
\author{
Benjamin Nutter
}
