% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprinkle_align.R
\name{sprinkle_align}
\alias{sprinkle_align}
\alias{sprinkle_align.default}
\alias{sprinkle_align.dust_list}
\title{Sprinkle Alignment of Table Cells}
\usage{
sprinkle_align(x, rows = NULL, cols = NULL, halign = NULL,
  valign = NULL, part = c("body", "head", "foot", "interfoot", "table"),
  fixed = FALSE, recycle = c("none", "rows", "cols", "columns"), ...)

\method{sprinkle_align}{default}(x, rows = NULL, cols = NULL,
  halign = NULL, valign = NULL, part = c("body", "head", "foot",
  "interfoot", "table"), fixed = FALSE, recycle = c("none", "rows", "cols",
  "columns"), ...)

\method{sprinkle_align}{dust_list}(x, rows = NULL, cols = NULL,
  halign = NULL, valign = NULL, part = c("body", "head", "foot",
  "interfoot", "table"), fixed = FALSE, recycle = c("none", "rows", "cols",
  "columns"), ...)
}
\arguments{
\item{x}{An object of class \code{dust}}

\item{rows}{Either a numeric vector of rows in the tabular object to be 
modified or an object of class \code{call}.  When a \code{call}, 
generated by \code{quote(expression)}, the expression resolves to 
a logical vector the same length as the number of rows in the table.
Sprinkles are applied to where the expression resolves to \code{TRUE}.}

\item{cols}{Either a numeric vector of columns in the tabular object to
be modified, or a character vector of column names. A mixture of 
character and numeric indices is permissible.}

\item{halign}{\code{character} One of \code{"left"}, \code{"center"}, 
or \code{"right"}. Defaults to \code{NULL}, for no change to the 
current value.}

\item{valign}{\code{character} One of \code{"top"}, \code{"middle"}, 
or \code{"bottom"}. Defaults to \code{NULL}, for no change to the 
current value.}

\item{part}{A character string denoting which part of the table to modify.}

\item{fixed}{\code{logical(1)} indicating if the values in \code{rows} 
and \code{cols} should be read as fixed coordinate pairs.  By default, 
sprinkles are applied at the intersection of \code{rows} and \code{cols}, 
meaning that the arguments do not have to share the same length.  
When \code{fixed = TRUE}, they must share the same length.}

\item{recycle}{A \code{character} one that determines how sprinkles are 
managed when the sprinkle input doesn't match the length of the region
to be sprinkled.  By default, recycling is turned off.  Recycling 
may be performed across rows first (left to right, top to bottom), 
or down columns first (top to bottom, left to right).}

\item{...}{Additional arguments to pass to other methods. Currently ignored.}
}
\description{
The alignment refers to the positioning of the text within
  a cell.  Alignment may be given relative to the left, center, or right
  of a cell, and the top, middle, or bottom of the cell.
}
\section{Functional Requirements}{

\enumerate{
 \item Correctly reassigns the appropriate elements of \code{halign} 
   and \code{valign} columns in the table part.
 \item Casts an error if \code{x} is not a \code{dust} object.
 \item Casts an error if \code{halign} is not a \code{character}
 \item Casts an error if \code{part} is not one of \code{"body"}, 
   \code{"head"}, \code{"foot"}, or \code{"interfoot"}
 \item Casts an error if \code{fixed} is not a \code{logical(1)}
 \item Casts an error if \code{recycle} is not one of \code{"none"},
   \code{"rows"}, or \code{"cols"}
 \item Casts an error if \code{valign} is not a \code{character}
 \item Cast an error if \code{recycle = "none"} and \code{halign} does
   not have length 1.
 \item Cast an error if \code{recycle = "none"} and \code{valign} does
   not have length 1.
 \item Cast an error if \code{halign} is not one of \code{c("left", "center", "right")}
 \item Cast an error if \code{valign} is not one of \code{c("top", "middle", "bottom")}
}

The functional behavior of the \code{fixed} and \code{recycle} arguments 
is not tested for this function. It is tested and validated in the
tests for \code{\link{index_to_sprinkle}}.
}

\seealso{
\code{\link{sprinkle}}, 
  \code{\link{index_to_sprinkle}}
}
