% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixiedust-pkg.R
\docType{data}
\name{sprinkle_documentation}
\alias{sprinkle_documentation}
\title{Sprinkle Documentation}
\format{
An object of class \code{data.frame} with 365 rows and 3 columns.
}
\usage{
sprinkle_documentation
}
\description{
Sprinkle Documentation
}
\author{
Benjamin Nutter \email{benjamin.nutter@gmail.com>}
}
\keyword{data}
