% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixiedust-pkg.R
\docType{data}
\name{sprinkle_reference}
\alias{sprinkle_reference}
\title{Sprinkle Reference}
\format{
An object of class \code{data.frame} with 47 rows and 9 columns.
}
\usage{
sprinkle_reference
}
\description{
Sprinkle Reference
}
\author{
Benjamin Nutter \email{benjamin.nutter@gmail.com>}
}
\keyword{data}
