% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixiedust-pkg.R
\docType{data}
\name{SprinkleRef}
\alias{SprinkleRef}
\title{SprinkleRef}
\format{
An object of class \code{data.frame} with 47 rows and 9 columns.
}
\usage{
SprinkleRef
}
\description{
SprinkleRef
}
\author{
Benjamin Nutter \email{benjamin.nutter@gmail.com>}
}
\keyword{data}
