% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixiedust-pkg.R
\docType{data}
\name{sprinkles}
\alias{sprinkles}
\title{Sprinkles}
\format{
An object of class \code{data.frame} with 46 rows and 6 columns.
}
\usage{
sprinkles
}
\description{
Sprinkles
}
\author{
Benjamin Nutter \email{benjamin.nutter@gmail.com>}
}
\keyword{data}
