% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_extract_base.R
\name{str_extract_base}
\alias{str_extract_base}
\alias{str_split_fixed_base}
\title{Extract Patterns from Character Strings}
\source{
https://stackoverflow.com/a/27274231/1017276
}
\usage{
str_extract_base(x, pattern)

str_split_fixed_base(x, pattern, n)
}
\arguments{
\item{x}{\code{character} vector.}

\item{pattern}{\code{character(1)} of the pattern to find in \code{x}}

\item{n}{The number of splits.}
}
\description{
This is a utility function that follow the pattern of
  \code{stringr::str_extract_all}.  It is provided to avoid the 
  dependency on the \code{stringr} package.
}
\seealso{
\code{stringr::str_extract_all}
}
