% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/table_attribute_bunnies.R
\name{table_attribute_bunnies}
\alias{dust_border_collapse}
\alias{dust_colnames}
\alias{dust_head_halign}
\alias{dust_table_border}
\alias{table_attribute_bunnies}
\title{Column Attributes for \code{dustpan} Tables}
\usage{
dust_colnames(...)

dust_head_halign(...)

dust_border_collapse(collapse = c("collapse", "separate"))

dust_table_border(sides, thickness = 1, units = c("px", "pt"),
  style = c("solid", "dashed", "dotted"), color = "black")
}
\arguments{
\item{...}{Column attributes for the table.  See 'Input Formats'}

\item{collapse}{Character string indicating if cell borders should blend together or
be distinct.}

\item{sides}{A character vector of up to length 4.  May use any of \code{"top"}, \code{"bottom"},
\code{"left"} or \code{"right"}.  The border style is applied to the sides of the table
specified.  Multiple sides are accepted and partial matching is performed.}

\item{thickness}{A numeric vector of length 1 specifying the thickness of the border.}

\item{units}{A character string indicating the units of \code{thickness}.}

\item{style}{A character string giving the style for the border line.  Only the first value is
accepted.}

\item{color}{A character string denoting the color of the border.  Only one value is permitted.}
}
\description{
Assigns new column attributes to a table
}
\section{Input Formats}{

\itemize{
  \item{\code{attr_vec}}{ Using \code{dust_colnames(attr_vec = c(...))}, a vector with named or
    unnamed elements may be provided.  If the elements are unnamed, it must have the same
    length as the number of columns in the table.}
  \item{named arguments}{ Using \code{dust_colnames(term = "Term", estimate = "Estimate")},
    column names may be passed for all or a subset of the columns.  The existing column
    name will be matched against the argument name.}
  \item{unnamed arguments}{ Using \code{dust_colnames("Term", "Estimate", "SE", ...)},
    column names may be passed for all of the columns.  If the arguments are unnamed, the
    number of arguments passed must match the number of columns in the table.}
 }
 When using named arguments (or a named vector), you may not mix named and unnamed elements.
 In other words, if one element is named, they must all be named.  Unnamed elements are assigned
 to columns in sequential order.
}
\examples{
x <- dust(lm(mpg ~ qsec + factor(am), data = mtcars))
x
x + dust_colnames(term = "Term", statistic = "T")
x + dust_colnames("Term", "Estimate", "SE", "T-statistic", "p-value")
\dontrun{
# Causes an error due to too few unnamed arguments
x + dust_colnames("Term", "Estimate")
}
}
\author{
Benjamin Nutter
}

