% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_bunnies.R
\name{add_bunnies}
\alias{+.dust}
\alias{add_bunnies}
\title{Modify Attributes of a \code{dustpan} table}
\usage{
\method{+}{dust}(x, y)
}
\arguments{
\item{x}{a \code{dust} object}

\item{y}{a \code{dust_bunny} object.}
}
\description{
Adds attributes of a plot to be rendered by the \code{print.dust} method.
}
\section{Dust Bunnies}{

The following is a list of available functions for adding dust bunnies to a \code{dust} table.
This list may or may not be complete.  Not every dust bunny will impact every type of table.
See \code{vignette("dustbunnies")} for a tabulation of which dustbunnies are applicable to each
printing method.

\itemize{
  \item{\code{\link{dust_cell_halign}}}{ Assign the horizontal alignment in a cell}
  \item{\code{\link{dust_colnames}}}{ Change the column names for a table}
  \item{\code{\link{dust_bold}}}{ Bold text in cells }
  \item{\code{\link{dust_fn}}}{ Apply a function to values in a table}
  \item{\code{\link{dust_head_halign}}}{ Assign the horizontal alignment for a column}
  \item{\code{\link{dust_italic}}}{ Italicize text in cells }
  \item{\code{\link{dust_round}}}{ Round values in cells }
}
}
\author{
Benjamin Nutter
}

