% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{\%<>\%}
\alias{\%<>\%}
\title{Chain together multiple operations.}
\usage{
lhs \%<>\% rhs
}
\arguments{
\item{lhs, rhs}{A dataset and function to apply to it}
}
\description{
This is a copy of the documentation for \code{\%<>\%} in 
\code{magrittr}.  The copy here is made to conform to CRAN requirements 
regarding documentation.  Please see the \code{magrittr} documenation
for the complete and current documentation.
}

