% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dust_colnames.R
\name{dust_colnames}
\alias{dust_colnames}
\title{Column Names for \code{dustpan} Tables}
\usage{
dust_colnames(...)
}
\arguments{
\item{...}{Column names for the table.  See 'Input Formats'}
}
\description{
Assigns new column names to a table
}
\section{Input Formats}{

\itemize{
  \item{\code{col_names}}{Using \code{dust_colnames(col_names = c(...))}, a vector with named or
    unnamed elements may be provided.  If the elements are unnamed, it must have the same
    length as the number of columns in the table.}
  \item{named arguments}{Using \code{dust_colnames(term = "Term", estimate = "Estimate")},
    column names may be passed for all or a subset of the columns.  The existing column
    name will be matched against the argument name.}
  \item{unnamed arguments}{Using \code{dust_colnames("Term", "Estimate", "SE", ...)},
    column names may be passed for all of the columns.  If the arguments are unnamed, the
    number of arguments passed must match the number of columns in the table.}
 }
 When using named arguments (or a named vector), you may not mix named and unnamed elements.
 In other words, if one element is named, they must all be named.  Unnamed elements are assigned
 to columns in sequential order.
}
\examples{
\dontrun{
# to use in tests
dust_colnames(col_names = c("Term", "Estimate", "SE", "T", "p"))
dust_colnames(term = c("Term", "SE"), std.error = "SE", p.value = "p")
dust_colnames(term = "Term", std.error = "SE", p.value = "p")
dust_colnames(term = "Term", "SE", "p")
dust_colnames("Term", "SE", "p")
}
}
\author{
Benjamin Nutter
}

